requirements.txt
telethon
tgcrypto

aiogram>=3.0

sqlalchemy
aiosqlite

numpy
sentence-transformers
torch --index-url https://download.pytorch.org/whl/nightly/cu128/

python-dotenv
